[FIX] web_id error for objects that have duplicated cards + order of existing objects in edit object panel.
The error raised when trying to call the web_id property on objects that have duplicates was AttributeError. It created confusion because getattr would then called __getattr__ which in turn calls getattr on the _modeling_object attribute. Hence the final error message was that the web_id attribute didn’t exist for the underlying efootprint object. That was fixed by turn the AttributeError into a PermissionError.